chore: split configurations into individual files
