+++ packages/middleware-signing/package.json
@@ -22,6 +22,7 @@
     "@aws-sdk/property-provider": "*",
     "@aws-sdk/protocol-http": "*",
     "@aws-sdk/signature-v4": "*",
+    "@aws-sdk/util-middleware": "*",
     "@aws-sdk/types": "*",
     "tslib": "^2.3.1"
   },
