+++ packages/middleware-signing/src/configurations.ts
@@ -1,18 +1,13 @@
-import { memoize } from "@aws-sdk/property-provider";
-import { SignatureV4, SignatureV4CryptoInit, SignatureV4Init } from "@aws-sdk/signature-v4";
+import { SignatureV4CryptoInit, SignatureV4Init } from "@aws-sdk/signature-v4";
 import {
   Credentials,
   HashConstructor,
   Logger,
   MemoizedProvider,
   Provider,
-  RegionInfo,
   RegionInfoProvider,
   RequestSigner,
 } from "@aws-sdk/types";
-
-// 5 minutes buffer time the refresh the credential before it really expires
-const CREDENTIAL_EXPIRE_WINDOW = 300000;
 
 // AwsAuth v/s SigV4Auth
 // AwsAuth: specific to SigV4 auth for AWS services
@@ -75,7 +70,7 @@ export interface SigV4AuthInputConfig {
   systemClockOffset?: number;
 }
 
-interface PreviouslyResolved {
+export interface AwsAuthPreviouslyResolved {
   credentialDefaultProvider: (input: any) => MemoizedProvider<Credentials>;
   region: string | Provider<string>;
   regionInfoProvider: RegionInfoProvider;
@@ -86,7 +81,7 @@ interface PreviouslyResolved {
   useDualstackEndpoint: Provider<boolean>;
 }
 
-interface SigV4PreviouslyResolved {
+export interface SigV4AuthPreviouslyResolved {
   credentialDefaultProvider: (input: any) => MemoizedProvider<Credentials>;
   region: string | Provider<string>;
   signingName: string;
@@ -116,115 +111,3 @@ export interface AwsAuthResolvedConfig {
 }
 
 export interface SigV4AuthResolvedConfig extends AwsAuthResolvedConfig {}
-
-export const resolveAwsAuthConfig = <T>(
-  input: T & AwsAuthInputConfig & PreviouslyResolved
-): T & AwsAuthResolvedConfig => {
-  const normalizedCreds = input.credentials
-    ? normalizeCredentialProvider(input.credentials)
-    : input.credentialDefaultProvider(input as any);
-  const { signingEscapePath = true, systemClockOffset = input.systemClockOffset || 0, sha256 } = input;
-  let signer: Provider<RequestSigner>;
-  if (input.signer) {
-    //if signer is supplied by user, normalize it to a function returning a promise for signer.
-    signer = normalizeProvider(input.signer);
-  } else {
-    //construct a provider inferring signing from region.
-    signer = () =>
-      normalizeProvider(input.region)()
-        .then(
-          async (region) =>
-            [
-              (await input.regionInfoProvider(region, {
-                useFipsEndpoint: await input.useFipsEndpoint(),
-                useDualstackEndpoint: await input.useDualstackEndpoint(),
-              })) || {},
-              region,
-            ] as [RegionInfo, string]
-        )
-        .then(([regionInfo, region]) => {
-          const { signingRegion, signingService } = regionInfo;
-          //update client's singing region and signing service config if they are resolved.
-          //signing region resolving order: user supplied signingRegion -> endpoints.json inferred region -> client region
-          input.signingRegion = input.signingRegion || signingRegion || region;
-          //signing name resolving order:
-          //user supplied signingName -> endpoints.json inferred (credential scope -> model arnNamespace) -> model service id
-          input.signingName = input.signingName || signingService || input.serviceId;
-
-          const params: SignatureV4Init & SignatureV4CryptoInit = {
-            ...input,
-            credentials: normalizedCreds,
-            region: input.signingRegion,
-            service: input.signingName,
-            sha256,
-            uriEscapePath: signingEscapePath,
-          };
-          const signerConstructor = input.signerConstructor || SignatureV4;
-          return new signerConstructor(params);
-        });
-  }
-
-  return {
-    ...input,
-    systemClockOffset,
-    signingEscapePath,
-    credentials: normalizedCreds,
-    signer,
-  };
-};
-
-// TODO: reduce code duplication
-export const resolveSigV4AuthConfig = <T>(
-  input: T & SigV4AuthInputConfig & SigV4PreviouslyResolved
-): T & SigV4AuthResolvedConfig => {
-  const normalizedCreds = input.credentials
-    ? normalizeCredentialProvider(input.credentials)
-    : input.credentialDefaultProvider(input as any);
-  const { signingEscapePath = true, systemClockOffset = input.systemClockOffset || 0, sha256 } = input;
-  let signer: Provider<RequestSigner>;
-  if (input.signer) {
-    //if signer is supplied by user, normalize it to a function returning a promise for signer.
-    signer = normalizeProvider(input.signer);
-  } else {
-    signer = normalizeProvider(
-      new SignatureV4({
-        credentials: normalizedCreds,
-        region: input.region,
-        service: input.signingName,
-        sha256,
-        uriEscapePath: signingEscapePath,
-      })
-    );
-  }
-
-  return {
-    ...input,
-    systemClockOffset,
-    signingEscapePath,
-    credentials: normalizedCreds,
-    signer,
-  };
-};
-
-const normalizeProvider = <T>(input: T | Provider<T>): Provider<T> => {
-  if (typeof input === "object") {
-    const promisified = Promise.resolve(input);
-    return () => promisified;
-  }
-  return input as Provider<T>;
-};
-
-const normalizeCredentialProvider = (
-  credentials: Credentials | Provider<Credentials>
-): MemoizedProvider<Credentials> => {
-  if (typeof credentials === "function") {
-    return memoize(
-      credentials,
-      (credentials) =>
-        credentials.expiration !== undefined &&
-        credentials.expiration.getTime() - Date.now() < CREDENTIAL_EXPIRE_WINDOW,
-      (credentials) => credentials.expiration !== undefined
-    );
-  }
-  return normalizeProvider(credentials);
-};
